FastApiBackend/products/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.95.2
fastapi-cli==0.0.5
h11==0.14.0
hiredis==2.4.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.14.0
pptree==3.1
pycparser==2.22
pydantic==1.10.7
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
python-ulid==1.1.0
PyYAML==6.0.2
redis==4.6.0
redis-om==0.1.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.27.0
typer==0.13.1
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==75.5.0.20241122
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
